+++ core/accounts/api/v1/serializers/__init__.py
@@ -1,0 +1,2 @@
+from .profiles import *
+from .accounts import *
