+++ core/accounts/api/v1/serializers/accounts.py
@@ -1,6 +1,5 @@
-from typing import Any, Dict
 from rest_framework import serializers
-from accounts.models import User, Profile
+from accounts.models import User
 from django.contrib.auth.password_validation import validate_password
 from django.core import exceptions
 from django.utils.translation import gettext_lazy as _
@@ -88,11 +87,3 @@ class ChangePasswordSerializer(serializers.Serializer):
             raise serializers.ValidationError({'new_password':list(e.messages)})           
         return super().validate(attrs)
   
-class ProfileSerializers(serializers.ModelSerializer):
-    email = serializers.CharField(source='user.email',read_only=True)
-    class Meta:
-        model = Profile
-        fields = ['id','email','first_name','last_name','image','descriptions']
-    
-        
-
