+++ core/accounts/api/v1/serializers/profiles.py
@@ -1,0 +1,12 @@
+from rest_framework import serializers
+from accounts.models import Profile
+
+
+class ProfileSerializers(serializers.ModelSerializer):
+    email = serializers.CharField(source='user.email',read_only=True)
+    class Meta:
+        model = Profile
+        fields = ['id','email','first_name','last_name','image','descriptions']
+    
+        
+
