+++ core/accounts/api/v1/urls/__init__.py
@@ -1,0 +1,7 @@
+from django.urls import path, include
+
+
+urlpatterns = [
+    path("",include('accounts.api.v1.urls.accounts')),
+    path("profile/",include('accounts.api.v1.urls.profiles')),
+]
