+++ core/accounts/api/v1/urls/profiles.py
@@ -1,0 +1,8 @@
+from django.urls import path
+from ..views import ProfileApiViews
+
+
+urlpatterns = [
+    # Profile
+    path('', ProfileApiViews.as_view(), name='profile'),
+]
